garm: introduce vm_location param
the current location (eastus) will only support SEV-SNP and nestable
instances, if we want to support TDX runners, we have to move to a
region that supports all three of them: westeurope is one of those
regions.

Since we don't want to invalidate all the urls in the github webhook
configurations when moving to another location (garm urls are bound
to a location), a discrete `vm_location` param has been introduced that
configures GARM to spawn Runners in that region.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>