Bump garm revision
This moves to a newer garm and garm-provider-azure revision that has been
tested and includes support for gallery images on Azure. This revision uses
"log/slog" so depends on being built with go 1.21.

Since the go 1.21 build container defaults to being bookworm based, specify
that explicitly and use a matching bookworm debian base image for the final
container.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>