garm: store github tokens in key vault
Instead of providing a list of name/token pairs as a deployment
variable, we will specify a key vault from which all secrets will be
read and converted from secret_name/value to a name/token pair.

The logic of converting the name/token to a garm config remains
untouched by this change.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>